fix: Supabase client build crash + hackathon auth-edge refactor
- Fix createClient() throwing during SSR/prerender when env vars unavailable
  (Cloudflare Pages build). Returns Proxy that defers error until actual use.
- Add singleton pattern to reuse browser Supabase client.
- Refactor hackathon edge routes to import from auth-edge instead of auth
  (separates Node.js crypto imports from edge runtime code).
- Remove stale ralph-loop.local.md state file.
- Update Supabase CLI temp version.
