--- .sisyphus/ralph-loop.local.md
@@ -1,30 +1,0 @@
----
-active: true
-iteration: 11
-max_iterations: 500
-completion_promise: "DONE"
-initial_completion_promise: "DONE"
-started_at: "2026-04-14T16:32:13.805Z"
-session_id: "ses_274ea3f77ffeHNjp3V45iM707p"
-ultrawork: true
-strategy: "continue"
-message_count_at_start: 96
----
-Exceeded free resources
-Build Minutes
-2h 6m / 0s
-Speed Insights Data Points
-53K / 10K
-Edge Requests
-1.6M / 1M
-10M
-Function Invocations
-1.4M / 1M
-Fast Origin Transfer
-7.71 GB / 10 GB
-100 GB
-Web Analytics Events
-36K / 50K
-100K
-Function Duration
-66.3 GB-Hrs / 100 GB-Hrs
