+++ app/api/hackathon/logout/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { SESSION_COOKIE } from "@/lib/hackathon/auth";
+import { SESSION_COOKIE } from "@/lib/hackathon/auth-edge";
 import { deleteSession } from "@/lib/hackathon/db";
 import { cookies } from "next/headers";
