+++ app/api/hackathon/me/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import { getSessionParticipant, updateParticipant } from "@/lib/hackathon/db";
 import { createClient } from "@/utils/supabase/server";
-import { getCorsHeaders, extractHackathonToken } from "@/lib/hackathon/auth";
+import { getCorsHeaders, extractHackathonToken } from "@/lib/hackathon/auth-edge";
 
 export const runtime = "edge";
