+++ app/api/hackathon/pre-questionnaire/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import { getSessionParticipant } from "@/lib/hackathon/db";
 import { createClient } from "@supabase/supabase-js";
-import { getCorsHeaders, extractHackathonToken } from "@/lib/hackathon/auth";
+import { getCorsHeaders, extractHackathonToken } from "@/lib/hackathon/auth-edge";
 
 function getAdminClient() {
   return createClient(
