+++ app/api/hackathon/push-subscribe/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { getSessionParticipant } from "@/lib/hackathon/db";
-import { getCorsHeaders, extractHackathonToken } from "@/lib/hackathon/auth";
+import { getCorsHeaders, extractHackathonToken } from "@/lib/hackathon/auth-edge";
 import { createClient as createServiceClient } from "@supabase/supabase-js";
 
 function getServiceClient() {
