+++ app/api/hackathon/register/special/route.ts
@@ -1,5 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
-import { hashPassword, generateSessionToken, SESSION_COOKIE, SESSION_EXPIRY_DAYS } from "@/lib/hackathon/auth";
+import { hashPassword, generateSessionToken } from "@/lib/hackathon/auth";
+import { SESSION_COOKIE, SESSION_EXPIRY_DAYS } from "@/lib/hackathon/auth-edge";
 import { findParticipantByEmail, createParticipant, createSession, getHackathonAdminClient } from "@/lib/hackathon/db";
 
 const SPECIAL_CAP = 0;
