+++ app/api/hackathon/team-matching/rankings/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { cookies } from "next/headers";
-import { SESSION_COOKIE } from "@/lib/hackathon/auth";
+import { SESSION_COOKIE } from "@/lib/hackathon/auth-edge";
 import {
   getHackathonMatchingEvent,
   getParticipantHackathonMatchingState,
