+++ app/api/hackathon/team/create/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-import { SESSION_COOKIE } from "@/lib/hackathon/auth";
+import { SESSION_COOKIE } from "@/lib/hackathon/auth-edge";
 import { getSessionParticipant, createTeam } from "@/lib/hackathon/db";
 import { cookies } from "next/headers";
