+++ app/api/hackathon/team/invite/status/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server";
 import { cookies } from "next/headers";
-import { SESSION_COOKIE } from "@/lib/hackathon/auth";
+import { SESSION_COOKIE } from "@/lib/hackathon/auth-edge";
 import { getSessionParticipant, getParticipantTeam } from "@/lib/hackathon/db";
 import { getExistingInvite, isInviteEnabled, isInvitedMember, teamAlreadyUsedInvite } from "@/lib/hackathon/invites";
