+++ app/api/hackathon/team/join/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-import { SESSION_COOKIE } from "@/lib/hackathon/auth";
+import { SESSION_COOKIE } from "@/lib/hackathon/auth-edge";
 import { getSessionParticipant, joinTeam } from "@/lib/hackathon/db";
 import { cookies } from "next/headers";
