+++ app/api/hackathon/team/match/cancel/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { SESSION_COOKIE } from "@/lib/hackathon/auth";
+import { SESSION_COOKIE } from "@/lib/hackathon/auth-edge";
 import { getSessionParticipant } from "@/lib/hackathon/db";
 import { cookies } from "next/headers";
 import { createClient } from "@/utils/supabase/server";
